ALL/requirements.txt
attrs==21.2.0
autopep8==1.5.7
click==8.0.1
decorator==5.0.9
flasgger==0.9.5
Flask==2.0.1
Flask-JWT-Extended==4.2.3
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
gunicorn==20.1.0
importlib-metadata==4.6.1
itsdangerous==2.0.1
Jinja2==3.0.1
jsonschema==3.2.0
MarkupSafe==2.0.1
mistune==0.8.4
pycodestyle==2.7.0
PyJWT==2.1.0
PyMySQL==1.0.2
pyrsistent==0.18.0
python-dotenv==0.18.0
PyYAML==6.0.1
six==1.16.0
SQLAlchemy==1.4.21
toml==0.10.2
typing-extensions==4.7.1
validators==0.18.2
Werkzeug==2.0.1
zipp==3.5.0
pypdf
openai
azure-functions
adlfs
langchain==0.0.254
langdetect
llama-index
azure-storage-blob 
python-docx
unstructured
azure-search-documents==11.4.0b4
openpyxl
langdetect
azure-identity
tiktoken
flask_cors==4.0.0
lingua-language-detector